Basic auth client names
