+++ README.md
@@ -308,7 +308,7 @@ To use form authentication in a web application, you should use the *org.pac4j.h
     HttpProfile httpProfile = client.getUserProfile(credentials, context);
     System.out.println("Hello: " + httpProfile.getUsername());</code></pre>
 
-To use basic auth authentication in a web application, you should use the *org.pac4j.http.client.indirect.BasicAuthClient* class the same way:
+To use basic auth authentication in a web application, you should use the *org.pac4j.http.client.indirect.IndirectBasicAuthClient* class the same way:
 
     // declare the client
     BasicAuthClient client = new BasicAuthClient(new MyUsernamePasswordAuthenticator(), new UsernameProfileCreator());
