+++ pac4j-http/src/main/java/org/pac4j/http/profile/HttpProfile.java
@@ -16,12 +16,12 @@
 package org.pac4j.http.profile;
 
 import org.pac4j.core.profile.CommonProfile;
-import org.pac4j.http.client.indirect.BasicAuthClient;
+import org.pac4j.http.client.indirect.IndirectBasicAuthClient;
 import org.pac4j.http.client.indirect.FormClient;
 
 /**
  * <p>This class is the user profile for sites using HTTP protocol.</p>
- * <p>It is returned by the {@link FormClient} or the {@link BasicAuthClient}.</p>
+ * <p>It is returned by the {@link FormClient} or the {@link IndirectBasicAuthClient}.</p>
  * <table summary="" border="1" cellspacing="2px">
  * <tr>
  * <th>Method :</th>
@@ -74,7 +74,7 @@ import org.pac4j.http.client.indirect.FormClient;
  * <p>All other attributes must be retrieved using the {@link #getAttributes()} method.</p>
  * 
  * @see FormClient
- * @see BasicAuthClient
+ * @see IndirectBasicAuthClient
  * @author Jerome Leleu
  * @since 1.4.0
  */
